split redusers from local and external data
