+++ src/index.jsx
@@ -1,16 +1,22 @@
-import {createStore, applyMiddleware} from "redux";
+import {createStore, applyMiddleware, combineReducers} from "redux";
 import {Provider} from "react-redux";
-import {reducer} from "./store/reducers/reducer";
 import React from "react";
 import ReactDOM from "react-dom";
 import thunk from "redux-thunk";
 import {compose} from "recompose";
 import App from "./components/app/app";
 import configureAPI from "./api";
-import Operation from './store/actions/async-actions';
+import Operation from "./store/actions/async-actions";
+import localData from './store/reducers/local-data/local-data';
+import appData from './store/reducers/app-data/app-data';
 
 const init = () => {
   const api = configureAPI((...args) => store.dispatch(...args));
+
+  const reducer = combineReducers({
+    localData,
+    appData
+  });
 
   const store = createStore(
       reducer,
