+++ src/store/reducers/app-data/app-data.js
@@ -1,5 +1,5 @@
 
-import {ActionType} from "../../actions/action-type.js";
+import {ActionType} from "../../actions/action-creator";
 
 const initialAppState = {
   questions: []
